user post deleted by admin api issue fix

+++ src/services/post.api.js
@@ -39,7 +39,7 @@ export const getSinglePost = async (id) => {
   return res.data;
 };
 
-export const deletePost = async (id) => {
-  const res = await API.delete(`/user/posts/${id}`);
+export const deletePost = async (id,role) => {
+  const res = await API.delete(`${user?.role_id == 2 ? "/user/posts" : "/admin/post"}/${id}`);
   return res.data;
 };
